Use pam for authentication instead of executing su
Executing su for authentication does not work universally as some versions
of it will refuse to run when not called on a tty. Using pam directly works
in all cases, but the programm needs the right to open /etc/shadow. Remote
authentication like nis or ldap and authenticating the own user usually
works without special rights.

This currently uses a fork of pam-auth, because upstream contains a bug that
prevents successfull authentication.