Record all API operations in a separate audit table
All actions that change some data are now recorded in a new audit_logs
table including the author, time, affected group (if any) and a textual
representation of the change that allows for a complete reconstruction
of the database row.

Existing columns in other tables (completions.{tutor,completed_at} and
elaborations.accepted_by) that previously served a similer purpose are
dropped and their data converted into the new format.