Only allow deleting students from empty groups
Students can no longer be removed from groups if the group has already
completed any tasks or handed in an elaboration.

As an alternative, the improved help message now suggests to create new
groups and mark the old one as no longer used by adding (ENDE) to the
comment field.