Add database table and admin interface to manage tutors
This adds a new mechanism for managing tutors and admins that is intended
to replace the existing one based on a list in the config file of users
that are allowed to log in.

This first commit adds the database table itself and an admin interface
for managing the entries. Each entry contains the username, the year it
is for, and a amind flag indicating if the user should be able to manage
the hardware course.