Use tutors database table to authenticate users
The config file now only contains site admins. The years for which a user
is a tutor or an admin are stored together with the username in a new cookie
named "user". If the user is a site admin, they are not bound by any year,
so only a flag indicating they are a site admin is stored.

This commit does not actually implement any access control for the site or
the api.