Implement access control for tutors, admins and site admins
Previously, every logged in user could change everything. This commit
introduces an access control schema with three roles that is enforced
on every http endpoint (sites and api). This builds upon the previous
two commits which added the necessary database changes and implemented
the configuration and login management changes.

Under the new schema, tutors and admins are constrained to a single
year, while site admins, which are still configured in the config file
and not through the database, can access all years.

The first role, tutor, can manage groups during the hardware course,
accept tasks and elaborations, and view the results of the students.
The admin can additionally help organize the hardware course itself by
managing experiments, tasks, events and students. The site admin can in
addition create and manage tutors and admins, can view the audit log,
and create new years.

All tasks and sites that are not available to the currently logged in
user are hidden.