[ff] Fix event not being deleted when clearing the date picker
Recent versions of firefox do not emit a change event when clearing a date
picker that has a default value and was not modified beforehand. To
workaround, we always delete the event when the date becomes empty in the
input event. This leads to duplicate delete calls when the change event
is actually emitted, so we have to make the api idempotent.