Push changes to other clients using server-sent events
This uses the hyper-sse crate to run a push server on a separate thread,
because Rocket does not support SSE yet.

Currently only changes to tasks, elaborations and comments are pushed,
but this can be easily extended. The year is used as a push channel
identifier, because it is the privilege boundary for tutors and is
complete in the sense that there is no page that displays specific
information that needs push messages from more than one year.

Alternatively, a set of groups could be used as a channel identifier.
This would theoretically result in less transferred messages, but as
all tutors are looking at the same or similar pages during a single
event of the hardware course, it would not make much practical
difference.