Add an optional IP address whitelist for tutors
The whitelist is disabled by default and does not affect site admins. It
can be configured through the admin interface.

It uses the postgresql inet support for storing hosts and nets and for
figuring out, if a host is part of a net. This works for both IPv4 and
IPv6.