Save given and family name separately
Previously the name of a student was stored as a single string. Even though
this is the correct way, it complicates the import and sorting, because the
original data source (almaweb) also stores the names separated. Before this
change, the name parts were concatenated before import, which is obviously
wrong. After this change, the names are imported and exported untransformed.

To still support the correct workflow and also cultures which do not fit into
this name schema, the family name is optional. The migration also just moves
the complete name into the given name field.

The display of the names is still incorrect in most of the tutor facing
views, as given and family name are still concatenated. However, the admin
view and export are now correct.